% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{calculate1}
\alias{calculate1}
\title{Calculate Mean, Variane, Standard Deviation}
\usage{
calculate1(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and stanard deviation of a vector
}
\examples{
calculate1(rnorm(10))
}
