% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{calculate2}
\alias{calculate2}
\title{Calculate Mean, Variane, SD (again)}
\usage{
calculate2(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector, but with user checks
}
\examples{
calculate2(rnorm(10))
}
