% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{calculate3}
\alias{calculate3}
\title{Compute Transpose And Inverse}
\usage{
calculate3(a, x)
}
\arguments{
\item{a}{matrix}

\item{x}{vector}
}
\value{
scalar
}
\description{
Given a numeric matrix A and a numeric vector `x`, 
calculates $x^T A^{-1} x$
Note that this only makes sense when A is a square matrix 
and the dimension of x is the same as the dimensions of the row and column dimensions of A
}
\examples{
x <- c(1,2,3,4)
mtx <-matrix(rnorm(16),nc=4,nr=4)
result8 <- calculate3(mtx, x)

}
