% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Example data from HW 3}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
data1
}
\description{
Example data from HW 3
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
