% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{gmlestimate}
\alias{gmlestimate}
\title{Maximum Likelihood Estimate with a gamma distribution}
\usage{
gmlestimate(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
Computes the maximum of the liklihood with a gamma distribution
}
\examples{
gmlestimate(rnorm(10))
}
