% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{hlcheck}
\alias{hlcheck}
\title{High-levels Check Function}
\usage{
hlcheck(x)
}
\arguments{
\item{x}{object}
}
\value{
object
}
\description{
Checks and throws error if not numeric, finit, zero lenth, NA, NAN
}
\examples{
hlcheck(NA)
}
