% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood Estimate of A Given Distribution for Data X}
\usage{
mle(x, func, interval)
}
\arguments{
\item{x}{vector}

\item{func}{function (experession)}

\item{interval}{vector(interval for optimize function)}
}
\value{
scalar
}
\description{
Estimate the maxium of liklihood of a given distribution for data x
}
\examples{
x1 = rgamma(100,3)
func1 = function(theta, x) dgamma(x, shape = theta, log = TRUE)
result <- mle(x1,func1,c(0,100))
}
