% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsmywapper.R
\name{mydplyr}
\alias{mydplyr}
\title{dplyr Wrapper}
\usage{
mydplyr(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
This is a wrapper for dplyr
}
\examples{
data(data1)
mydplyr(data1)
}
