% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsmywapper.R
\name{plotArea}
\alias{plotArea}
\title{Wrapper function for ggplot2 for data data1}
\usage{
plotArea(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance andata1 sdata1 of a vector
}
\examples{
data(data1)
plotArea(data1)
}
