% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{stdization}
\alias{stdization}
\title{Standardization}
\usage{
stdization(m)
}
\arguments{
\item{m}{matrix which has more than one row}
}
\value{
matrix
}
\description{
Take a numeric matrix and standardizes its columns
}
\examples{
m <-matrix(rnorm(16),nc=4,nr=4)
result9 <- stdization(m)

}
