% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{wcalculate1}
\alias{wcalculate1}
\title{Weighted Mean, Variane, and Standard Deviation}
\usage{
wcalculate1(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
list
}
\description{
Computes the weighted mean, variane, standard deviation
}
\examples{
d <- read.table(url("http://www.stat.umn.edu/geyer/3701/data/q1p4.txt"),header = TRUE)
data(d)
wcalculate1(d)
}
