% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangbaiLiToolsRfunctions.R
\name{wcalculate2}
\alias{wcalculate2}
\title{Weighted Mean, Variane, and Standard Deviation with User Checkes}
\usage{
wcalculate2(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
list
}
\description{
Computes the weighted Mean, Variane, and Standard Deviation with user checks
}
\examples{
d <- read.table(url("http://www.stat.umn.edu/geyer/3701/data/q1p4.txt"),header = TRUE)
data(d)
wcalculate2(d)
}
